fix controller

+++ routes/web.php
@@ -11,6 +11,8 @@ use App\Http\Controllers\AuthController;
 use App\Http\Controllers\AdminController;
 use App\Http\Controllers\PeminjamController;
 use App\Http\Controllers\ForgotPasswordController;
+use App\Http\Controllers\HistoriController;
+use App\Http\Controllers\TransaksiPeminjamanController;
 use App\Models\Peminjam;
 
 Route::get('/', function () {
@@ -19,23 +21,19 @@ Route::get('/', function () {
 
 
 
-// Route untuk menampilkan form login
+// login
 Route::get('/login', [AuthController::class, 'showLoginForm'])->name('login')->middleware('guest');
+Route::post('/login', [AuthController::class, 'login'])->name('login.process')->middleware('guest');
 
-// Route untuk menampilkan form register
+// register
 Route::get('/register', [AuthController::class, 'showRegisterForm'])->name('register')->middleware('guest');
-
-// Route untuk memproses login
-Route::post('/login', [AuthController::class, 'login'])->name('login.process')->middleware('guest');;
-
-// Route untuk memproses register
 Route::post('/register', [PeminjamController::class, 'create'])->name('register.process')->middleware('guest');;
+
+//Route untuk Logout
+Route::post('/logout', [AuthController::class, 'logout'])->name('logout')->middleware('auth');
 
 // route verifikasi email
 Route::get('/verify-email/{token}', [PeminjamController::class, 'verifyEmail'])->name('verify.email')->middleware('guest');
-
-//Route untuk Logout
-Route::post('/logout', [AuthController::class, 'logout'])->name('logout');
 
 // forgot password dan reset password
 Route::get('/forgot-password', [ForgotPasswordController::class, 'showForgotPasswordForm'])->name('password.request');
@@ -45,90 +43,91 @@ Route::get('/reset-password/{token}', [ForgotPasswordController::class, 'showRes
 Route::post('/reset-password', [ForgotPasswordController::class, 'resetPassword'])->name('password.update');
 
 
+
+
+
 // USER
-// Dashboard
-Route::get('/user/dashboard', [PeminjamController::class, 'index'])->middleware('auth')->name('user.dashboard');
+Route::get('/user/dashboard', [PeminjamController::class, 'userDashboard'])->middleware('auth')->name('user.dashboard');
 
 Route::get('/user/profile', [PeminjamController::class, 'userProfile'])->middleware('auth')->name('user.profile');
 
-// Peminjaman
+//user peminjaman
 Route::get('/user/peminjaman', [PeminjamController::class, 'userPeminjaman'])->middleware('auth')->name('user.peminjaman');
 
-Route::post('/user/peminjaman', [PeminjamController::class, 'Pinjam']);
+Route::post('/user/peminjaman', [TransaksiPeminjamanController::class, 'create'])->middleware('auth');
 
 Route::put('/user/{peminjam}/updateProfile', [PeminjamController::class, 'Update'])->middleware('auth')->name('user.update');
 
-// history
+//user history
 Route::get('/user/history', [PeminjamController::class, 'userHistory'])->middleware('auth')->name('user.history');
 
-//detail history
 Route::get('/user/detail/{id}', [peminjamController::class, 'userdetail'])->middleware('auth')->name('user.detail');
 
+
+
+
+
+
 //admin
-//Dashboard
-Route::get('/admin/dashboard', [AdminController::class, 'admindashboard'])->middleware('admin')->name('admin.dashboard');
+Route::get('/admin/dashboard', [AdminController::class, 'adminDashboard'])->middleware('admin')->name('admin.dashboard');
 
-Route::get('/admin/kelola', [AdminController::class, 'kelolapeminjaman'])->middleware('admin')->name('admin.kelola');
+Route::get('/admin/kelola', [AdminController::class, 'adminKelola'])->middleware('admin')->name('admin.kelola');
 
-Route::get('/admin/histori', [AdminController::class, 'historyadmin'])->middleware('admin')->name('admin.history');
+Route::get('/admin/histori', [AdminController::class, 'adminHistori'])->middleware('admin')->name('admin.history');
 
-Route::get('/admin/lanjutan', [AdminController::class, 'lanjutan'])->middleware('admin')->name('admin.lanjutan');
+Route::get('/admin/lanjutan', [AdminController::class, 'adminLanjutan'])->middleware('admin')->name('admin.lanjutan');
 
+Route::get('/admin/useradmin', [AdminController::class, 'adminUsers'])->middleware('admin')->name('admin.useradmin');
 
-Route::get('/admin/useradmin', [AdminController::class, 'useradmin'])->middleware('admin')->name('admin.useradmin');
-
-
-
-
-
+Route::get('/admin/detail/{id}', [AdminController::class, 'adminDetail'])->middleware('admin')->name('admindetail');
 
 Route::get('/admin/lanjutan/{id}', [AdminController::class, 'datalanjutan'])->middleware('admin')->name('adminlanjut');
 
-Route::get('/admin/terima/{id}', [AdminController::class, 'terimaStatus']);
 
-Route::post('/admin/tolak/{peminjam}', [AdminController::class, 'tolakStatus']);
 
-Route::delete('/admin/hapusUser/{peminjam}', [PeminjamController::class, 'delete']);
+//admin user
+Route::post('/admin/tolak/{peminjam}', [AdminController::class, 'tolakStatusUser'])->middleware('admin');
 
-Route::put('/admin/updateUser/{peminjam}', [AdminController::class, 'updateUser']);
+Route::get('/admin/terima/{id}', [AdminController::class, 'terimaStatusUser'])->middleware('admin');
 
-Route::post('/admin/kelola/simpan-ke-history/{transaksi}', [AdminController::class, 'simpanKeHistory'])->name('simpan-ke-history');
+Route::delete('/admin/hapusUser/{peminjam}', [PeminjamController::class, 'delete'])->middleware('admin');
 
-Route::get('/admin/detail/{id}', [AdminController::class, 'datadetail'])->middleware('admin')->name('admindetail');
+Route::put('/admin/updateUser/{peminjam}', [AdminController::class, 'updateUser'])->middleware('admin');
+
+Route::post('/admin/kelola/simpan-ke-history/{transaksi}', [HistoriController::class, 'simpanKeHistory'])->middleware('admin')->name('simpan-ke-history');
+
+
 
 
 // untuk imb
-Route::get('/admin/imb', [AdminController::class, 'manajemenImb'])->middleware('admin')->name('admin.manajemenImb');
+Route::get('/admin/imb', [ImbController::class, 'manajemenImb'])->middleware('admin')->name('admin.manajemenImb');
 
-Route::get('/admin/tambahImb', [AdminController::class, 'viewTambahImb'])->middleware('admin')->name('admin.viewTambahImb');
-Route::post('/admin/tambahImb', [AdminController::class, 'tambahImb'])->middleware('admin')->name('admin.tambahImb');
+Route::get('/admin/tambahImb', [ImbController::class, 'viewTambahImb'])->middleware('admin')->name('admin.viewTambahImb');
+
+Route::post('/admin/tambahImb', [ImbController::class, 'tambahImb'])->middleware('admin')->name('admin.tambahImb');
 // lihat file imb
-Route::get('/admin/lihat/imb/{name}', [AdminController::class, 'show'])->middleware('admin')->name('admin.lihatImb');
+Route::get('/admin/lihat/imb/{name}', [ImbController::class, 'lihatImb'])->middleware('admin')->name('admin.lihatImb');
 // edit imb
-Route::put('/admin/edit/imb/{id}', [AdminController::class, 'updateImb'])->middleware('admin')->name('edit.imb');
+Route::put('/admin/edit/imb/{id}', [ImbController::class, 'updateImb'])->middleware('admin')->name('edit.imb');
 // delete imb
-Route::delete('/admin/delete/imb/{id}', [AdminController::class, 'deleteImb'])->middleware('admin')->name('delete.imb');
-
+Route::delete('/admin/delete/imb/{id}', [ImbController::class, 'deleteImb'])->middleware('admin')->name('delete.imb');
 // search
 Route::get('/admin/imb/search', [ImbController::class, 'search'])->middleware('admin')->name('search.imb');
 // print all
 Route::get('/admin/imb/printAll', [ImbController::class, 'printAll'])->middleware('admin')->name('imb.printAll');
 
-Route::post('admin/kelola/{id}', [AdminController::class, 'konfirmasiPengembalian'])->name('konfirmasi.pengembalian');
 
-Route::post('admin/dashboard/{id}', [AdminController::class, 'Pengembalian'])->name('pengembalian.konfirmasi');
 
-Route::post('admin/pegambilan/{id}', [AdminController::class, 'pengambilan'])->name('pengambilan');
-
-Route::post('admin/batalkan/{id}', [AdminController::class, 'pembatalan'])->name('pembatalan');
 
 
 //untuk SK
 Route::get('/admin/sk', [SKController::class, 'manajemenSK'])->middleware('admin')->name('admin.manajemenSK');
+
 Route::get('/admin/tambahSK', [SKController::class, 'viewTambahSK'])->middleware('admin')->name('admin.viewTambahSK');
+
 Route::post('/admin/tambahSK', [SKController::class, 'tambahSK'])->middleware('admin')->name('admin.tambahSK');
 
-Route::get('/admin/lihat/sk/{name}', [SKController::class, 'show'])->middleware('admin')->name('admin.lihatSK');
+Route::get('/admin/lihat/sk/{name}', [SKController::class, 'lihatSk'])->middleware('admin')->name('admin.lihatSK');
 // edit imb
 Route::put('/admin/edit/sk/{id}', [SKController::class, 'updateSK'])->middleware('admin')->name('edit.sk');
 // hapus sk
@@ -139,19 +138,27 @@ Route::get('/admin/sk/search', [SKController::class, 'search'])->middleware('adm
 Route::get('/admin/sk/printAll', [SKController::class, 'printAll'])->middleware('admin')->name('sk.printAll');
 
 
-// untuk Keuangan
-Route::get('/admin/tambahKeuangan', [AdminController::class, 'viewTambahKeuangan'])->middleware('admin')->name('admin.viewTambahKeuangan');
-ROute::post('/admin/tambahKeuangan', [AdminController::class, 'tambahKeuangan'])->middleware('admin')->name('admin.tambahKeuangan');
+
+//untuk histori
+Route::post('admin/pengembalian/{id}', [HistoriController::class, 'pengembalian'])->name('konfirmasi.pengembalian');
+
+Route::post('admin/pegambilan/{id}', [HistoriController::class, 'pengambilan'])->name('pengambilan');
+
+Route::post('admin/batalkan/{id}', [HistoriController::class, 'pembatalan'])->name('pembatalan');
+
+
+
+
+//autocomplete search transaksi peminjaman
+Route::get('cari', [TransaksiPeminjamanController::class, 'autocomplete']);
+
 
 
 Route::get('/tes', function () {
     return view('tes');
 });
 
-Route::get('cari', [AdminController::class, 'autocomplete']);
-
-
-// testing
-Route::get('/test-reminder-emails', [PeminjamController::class, 'testSendReminderEmails']);
+// reminder email
+Route::get('/test-reminder-emails', [PeminjamController::class, 'sendReminderEmails']);
 
 // Route::get('/admin/kelola', [AdminController::class, 'kelolapeminjaman'])->middleware('admin')->name('admin.kelola');
